requirements.txt
# python package requirements file

# Note to our Learners: Our courses are created at a specific point in time. The models and libraries used will naturally change over time.
# If you download and run the notebooks locally, you should create a Python virtual environment  
# with the Python version matching the one used in the course and use this requirements.txt file to install the required Python packages.
# Due to the wide diversity of platforms and environments, DeepLearning does not explicitly support running locally. However, if you encounter
# issues, the course support channel is a good place to ask questions.


# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision

#for example:
#torch==2.0.1
#matplotlib==3.7.2

jupyter==1.1.1
colpali-engine[interpretability] @ git+https://github.com/kacperlukawski/colpali.git@feature/colsmol-patches
transformers==4.53.1
torch==2.5.1
torchvision==0.20.1
scikit-learn==1.7.1
plotly==6.3.0
qdrant-client==1.15.1
fastembed==0.7.3
pandas==2.3.2
pdf2image==1.17.0
pyarrow==21.0.0
openai==2.2.0
python-dotenv==1.0.1
streamlit==1.32.2
uvicorn
python-multipart
